requirements.txt
altgraph==0.17.3
appdirs==1.4.4
argcomplete==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
future==0.18.2
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
libusb==1.0.26b5
MarkupSafe==2.1.1
packaging==21.3
pefile==2022.5.30
Pillow==9.2.0
pkg-about==1.0.7
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
PyJWT==2.5.0
pyngrok==5.1.0
pyparsing==3.0.9
pyserial==3.5
PySimpleGUI==4.60.3
python-dotenv==0.21.0
python-escpos==3.0a7
pytz==2022.4
pyusb==1.2.1
pywin32-ctypes==0.2.0
PyYAML==6.0
qrcode==7.3.1
requests==2.28.1
six==1.16.0
tomli==2.0.1
twilio==7.14.2
urllib3==1.26.12
viivakoodi==0.8.0
Werkzeug==2.2.2
zipp==3.8.1
